Update to .NET Core
Update to .NET Core, .NET Standard and xUnit
