+++ CSharp/parrot.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28010.2003
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29326.143
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "parrot", "parrot.csproj", "{B92D4245-86CA-4380-BAEC-5560B749C31F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Parrot.Tests", "Parrot.Tests\Parrot.Tests.csproj", "{FC9C4B6B-C183-48D8-BF21-149CBE874013}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Parrot", "Parrot\Parrot.csproj", "{6F530397-2988-4435-A61B-967BE19AFD1C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +13,19 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B92D4245-86CA-4380-BAEC-5560B749C31F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B92D4245-86CA-4380-BAEC-5560B749C31F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B92D4245-86CA-4380-BAEC-5560B749C31F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B92D4245-86CA-4380-BAEC-5560B749C31F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FC9C4B6B-C183-48D8-BF21-149CBE874013}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FC9C4B6B-C183-48D8-BF21-149CBE874013}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FC9C4B6B-C183-48D8-BF21-149CBE874013}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FC9C4B6B-C183-48D8-BF21-149CBE874013}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6F530397-2988-4435-A61B-967BE19AFD1C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6F530397-2988-4435-A61B-967BE19AFD1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6F530397-2988-4435-A61B-967BE19AFD1C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6F530397-2988-4435-A61B-967BE19AFD1C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {EC9641C9-A896-4EC5-A986-6756FA955887}
+		SolutionGuid = {49B70617-A81F-459A-8EDA-52ABF95D02D0}
 	EndGlobalSection
 EndGlobal
